requirements.txt
pytest
requests
python-dotenv
gradio>=4.0.0
mutagen
